created a parser helper for reusable methods
